# api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Usage Dashboard API
  description: >-
    Contract for the AI Usage Dashboard `/api/v1` HTTP surface. This document is the
    single source of truth for the provider registry endpoints; types are generated
    from it via oapi-codegen and hand-written handlers are validated against it.
  version: "1.0.0"
servers:
  - url: /api/v1
tags:
  - name: providers
    description: Provider registry — metadata and enable/disable state.
paths:
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List all known providers with their enabled state.
      responses:
        "200":
          description: The full set of registered providers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
        "500":
          $ref: "#/components/responses/InternalError"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Get a single provider by id.
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: The requested provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /providers/{id}/enable:
    post:
      operationId: enableProvider
      tags: [providers]
      summary: Enable a provider. Idempotent — enabling an already-enabled provider succeeds.
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: The provider, now enabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /providers/{id}/disable:
    post:
      operationId: disableProvider
      tags: [providers]
      summary: Disable a provider. Idempotent — disabling an already-disabled provider succeeds.
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: The provider, now disabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /providers/{id}/credentials:
    put:
      operationId: setProviderCredentials
      tags: [providers]
      summary: >-
        Set every credential value for a provider (full replace). The request
        must supply exactly the provider's declared credential fields —
        unknown or missing fields are rejected. Values are encrypted at rest
        and are never returned by any read path.
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialValues"
      responses:
        "204":
          description: Credentials stored.
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: getProviderCredentials
      tags: [providers]
      summary: >-
        Get presence-only credential status for a provider. Never returns
        secret material or a masked hint — only whether each declared field
        currently has a stored value.
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: Presence of each of the provider's declared credential fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialPresenceList"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteProviderCredentials
      tags: [providers]
      summary: Clear all stored credential values for a provider.
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "204":
          description: Credentials cleared.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /providers/{id}/usage:
    get:
      operationId: getProviderUsage
      tags: [providers]
      summary: >-
        Get the latest usage snapshot for a provider. Latest-only — no
        historical data is retained; each snapshot replaces the previous one.
        A provider that has never been collected returns 200 with a null
        collectedAt and no metrics (pending state), not 404 — 404 is reserved
        for an unknown provider id.
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: The latest usage snapshot for the provider, or a pending placeholder.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsageSnapshot"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    ProviderId:
      name: id
      in: path
      required: true
      description: Stable provider identifier (e.g. `openai`, `anthropic`).
      schema:
        type: string
  schemas:
    CredentialField:
      type: object
      description: Metadata describing one credential field a provider requires.
      required: [name, label, secret]
      properties:
        name:
          type: string
          description: Machine-readable field key (e.g. `api_key`).
        label:
          type: string
          description: Human-readable label for display in a credential form.
        secret:
          type: boolean
          description: Whether the field value must be masked/encrypted at rest.
    Provider:
      type: object
      required: [id, name, enabled, credentialFields]
      properties:
        id:
          type: string
          description: Stable provider identifier.
        name:
          type: string
          description: Human-readable provider name.
        enabled:
          type: boolean
          description: Whether the provider is currently enabled.
        credentialFields:
          type: array
          description: Credential fields required to configure this provider.
          items:
            $ref: "#/components/schemas/CredentialField"
    CredentialValues:
      type: object
      description: >-
        Full set of credential values for a provider, keyed by declared
        field name. Write-only: this payload is never echoed back by any
        read path.
      required: [values]
      properties:
        values:
          type: object
          description: >-
            Field name → secret value. Must match the provider's declared
            credential fields exactly — unknown or missing keys are
            rejected.
          additionalProperties:
            type: string
    CredentialPresence:
      type: object
      description: Whether a declared credential field currently has a stored value.
      required: [name, configured]
      properties:
        name:
          type: string
          description: Machine-readable field key (e.g. `api_key`).
        configured:
          type: boolean
          description: Whether a value is currently stored for this field.
    CredentialPresenceList:
      type: object
      required: [fields]
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/CredentialPresence"
    UsageMetric:
      type: object
      description: One usage data point captured in a provider's usage snapshot.
      required: [name, window, unit, used]
      properties:
        name:
          type: string
          description: Metric name (e.g. `requests`, `spend`).
        window:
          type: string
          description: The window the metric covers (e.g. `day`, `month`).
        unit:
          type: string
          description: >-
            Smallest integer unit the metric is expressed in (e.g.
            `usd_cents`, `tokens`) — no fractional/float quantities.
        used:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
          nullable: true
          description: The provider's disclosed cap, or null if unlimited/unknown.
        remaining:
          type: integer
          format: int64
          nullable: true
          description: Provider-supplied remaining amount, or null if unknown.
        resetAt:
          type: string
          format: date-time
          nullable: true
          description: When this metric's window resets, or null if not disclosed.
    UsageSnapshot:
      type: object
      description: >-
        The latest-only usage snapshot for a provider. No history/time-series
        is retained — each collection replaces the previous snapshot.
      required: [collectedAt, metrics]
      properties:
        collectedAt:
          type: string
          format: date-time
          nullable: true
          description: >-
            When this snapshot was collected, or null if the provider has
            never been collected (pending state).
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/UsageMetric"
        lastError:
          type: string
          nullable: true
          description: >-
            Error message from the most recent failed collection attempt, if
            any; absent/null otherwise.
    Error:
      type: object
      description: Canonical error envelope returned by every `/api/v1` error response.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
              enum:
                - validation_error
                - not_found
                - conflict
                - unsupported_media_type
                - internal_error
            message:
              type: string
              description: Human-readable error message. Never contains internal detail on 5xx.
            details:
              type: object
              description: Optional structured detail (e.g. field-level validation errors).
              additionalProperties: true
  responses:
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: An unexpected server error occurred. Detail is never leaked to clients.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: The request failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: The request's Content-Type is not supported.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"